backend/services/editor-snapshot/bin/worker/requirements.txt
pulsar-client==3.11.0
asyncpg==0.29.0
aioboto3==12.4.0
dotenv
tenacity
aiofiles
aiohttp
uuid6
